rename jax to jx and add requirement-extra.txt
