+++ mars/tensor/fuse/__init__.py
@@ -17,4 +17,4 @@
 from .core import TensorFuseChunk
 from .ne import TensorNeFuseChunk
 from .cp import TensorCpFuseChunk
-from .jax import TensorJaxFuseChunk
+from .jx import TensorJaxFuseChunk
