+++ requirements-extra.txt
@@ -3,6 +3,8 @@ scipy>=1.0.0,<1.3.0; python_version < '3'
 scikit-learn>=0.20; python_version >='3'
 scikit-learn>=0.20,<0.21.0; python_version < '3'
 numexpr>=2.6.4
+jax>=0.1.41;sys_platform != 'win32'
+jaxlib>=0.1.22;sys_platform != 'win32'
 psutil>=4.0.0
 lz4>=1.0.0
 protobuf>=3.6
